[code-projects] remove obsolete "visible_rows_count" and "visible_columns_count" from "TableStructure"
In the early days of the plugin the properties "rows_count" and "columns_count" were used. With the displaying of only non-hidden (visible) rows and columns, these properties were no longer used.
They were left in place because it was not yet clear if the concept of displaying the unhidden data would work properly.

Since we only display data that is not hidden, the additional "visible" in the variable name is no longer necessary.