[plugin] remove trailing spaces from test data
Jsoup adds extra spaces at the end of lines (github jsoup #1689).

This fix is required because using jsoup for formatting the generated test data lead to different results when exporting via PyCharm debugger or the "generateTestData" gradle-tasks. Even if the output from the PyCharm debugger was the same as from the gradle-tasks.