[code-projects] pandas 1.3 - [cleanup] html props validation
We are only interested in whether the header rows and cells of the table have the same values and the same appearance. All other information contained in the properties are not relevant.

It is also not necessary that the indices of the chunks are all unique and match the expected result. For simplicity, these are removed completely instead of adjusting them.