[plugin] disable Ctrl+A table action
The focus border can be in a different column than the selected column. This could confuse the user when wanting to sort the current selected column.

Example:
- select first table cell -> focus is in first column
- Ctrl+A -> last column is now selected (single column selection) but the focused cell is in the first column

The action is disabled for the time being, because it is currently not possible to copy the values of the selected cells.

To make a component ignore a key that it normally responds to, one can use the special action name "none". For example, the following code makes a component ignore the F2 key.

component.getInputMap().put(KeyStroke.getKeyStroke("F2"), "none");

To also prevent searching in the other input maps, the "DoNothingAction" is registered instead.