[plugin] fix false positive (validation)
The validation reported in some rare cases a false positive.

To trigger the problem, a filter must be defined, or some rows/columns must be set to hidden, additionally the DataFrame must be smaller or equal to the size of the fetched chunk.

The code created a copy of the pandas Styler without the hidden rows/columns and used that copy to produce the expected result.