[docs] remove note about broken code completion
The non-working code completion happens only in my installed IntelliJ instance. I couldn't reproduce the behavior with another Jetbrains IDE.

The problem is not present when testing the plugin:
- in an IDE started via gradle task "rundIde"
- in PyCharm started from IntelliJ source code
- in a fresh installed PyCharm CE (latest for Mac)
- in a fresh installed IntelliJ CE (latest for Mac)

The IDE which causes the problem had the same "PyReferenceResolveProvider" registered in exact same order as when started from the plugin project (gradle task "runIde"). This was verified by logging the registered providers.