[plugin] use importlib.util.find_spec to check if pandas is available
Importing pandas takes some time. Therefore, it is not a good fit to check if pandas is available. importlib.util.find_spec is way faster because it doesn't import pandas. Since the check for pandas is executed for every paused debug session it should be as fast as possible to reduce side effects for the plugin users.