[plugin] fix: variables inside a function/method are not resolved
"PyResolveUtil.resolveQualifiedNameInScope" couldn't resolve variables defined in a function. If the same name was used for a global var it resolved to the global one.

I can't tell if it was used wrong or if it can't work for our use case. Switching back to the previous used "PyResolveUtil.resolveLocally" fixes the problem.

Example:

import pandas as pd

df2 = "abc"

def non_working_example():
    df2 = pd.DataFrame.from_dict({
        "col_0": [0, 1],
        "col_1": [2, 3],
    })

    breakpoint()

non_working_example()