[plugin] create temporary tableSource if executed in console
The console based debug process can't evaluate code like in a normal debug process. Evaluated expressions get no temp identifier assigned to refer to the result. Instead, the expression is assigned. Referring to the created value by using the expression, executes the expression again. Storing the result with a custom identifier in the Python stack frame makes it visible in the variable view of the debug console. To circumvent this limitation, the created variables are stored in a dict and are referenced via the dict.