[code-projects] improve Styler::background_gradient validation performance (part 1)
Validation for data frames with many columns and many rows is very slow. The following actions require most of the computing time:

- applying the styler "subset"
- computing the "gmap", "vmin" and "vmax"

The color gradient can only be applied to numerical values. When calling Styler::background_gradient without "subset", Pandas creates a list of boolean flags for all columns and uses them as "subset" internally. During the validation phase, the "subset" was used several times to create different sub-chunks to validate. The "subset" is now applied only once and the resulting DataFrame is cached. As the problem of multiple calculation occurs with all styling functions, the optimization is applied to all patchable styling functions.

In addition, the values "gmap", "vmin" and "vmax" of Styler::background_gradient are recalculated for each sub-chunk during the validation. The values themselves are independent of the sub-chunk to be calculated. To further improve performance, these values are now calculated once and cached.

These changes drastically improve performance, but also increase memory consumption.

For pandas < 1.3 "gmap" is calculated inside the styling function and can't therefore be cached. The improvements are relative small.