[code-projects] improve Styler::background_gradient validation performance (part 2)
During the validation a chunk is split up in smaller sub-chunks. A styling function is applied to the chunk and all the sub-chunks to compute the styling of these chunks. To improve performance, the original DataFrame is replaced by the chunk to be validated during validation.

This brings the following improvements:

1) reduced slice time
Slicing a small DataFrame is faster than slicing a large DataFrame.

A styler applies the "subset" of a styling function to extract the DataFrame to which the styling function is to be applied. Extraction takes place when calculating the styling for the chunk and for each sub-chunk.

The saving therefore applies to every chunk and sub-chunk.

2) reduced computation (general, no improvement for Styler::background_gradient)
Styling function which use the "chunk_parent" to compute information like max/min value result in a slow validation for large DataFrames. Because, the min/max value is calculated multiple times (for the chunk and each sub-chunk).

While the correct maximum value is required to determine a visual result of a chunk, it is not important for validation as this data is not displayed.