[code-projects] use enum for 'CreateTableSourceFailure::error_kind' instead of string
On Kotlin side the string is mapped to an enum. To not break the deserializing of the JSON value, an enum on Python side with same field names is used instead of a string.

The JSON formatter used on Python side was adjusted to serialize the name of an enum instead of "<class_name>.<name>".