[plugin] unlink source of tableSource on dispose
Nullifying of data (added in the previous commit) seem to work on linux but not on macOS.
Therefore, the temporary variables, created by the PyDev debugger stay in memory after the user has closed the viewer (dialog) as long as the current stack frame is not dropped. This could lead to increased lifetime of the viewed DataFrame. To prevent this, the internal references to the viewed DataFrame are nullified in an additional step.