[plugin] lazy loading of column statistics/use colNames from tableStructure
Statistics for a DataFrame column are now fetched when showing the column tooltip. Previously this data was included for each column of the chunk, if they were not loaded before.

Resorting a DataFrame and fetching the first chunk of the re-sorted DataFrame is now done in one step. Previously the sorting was first applied and then an extra call was made to fetch the data of the sorted DataFrame.

Both adjustments should slightly reduce the loading time for querying the data from the DataFrame.

Column names are taken from the TableStructure. Therefore, these data is dropped from the chunk values.