[plugin] Introduce 'ILoadRequestCreator'
The load request creator is used to specify which data should be loaded for a chunk at the time the load request is executed. This ensures that optional data is not loaded multiple times.

Example of such a case:

Row index names are optional in a chunk. If they have already been loaded, they do not need to be loaded a second time when the data of other columns of the same rows should be fetched.

If the row index labels are not yet loaded and the user is scrolling horizontally, the table model requests multiple chunk regions. All of which would request the row index labels until one of the loaded chunks provides these labels. The other already requested regions would also retrieve this optional data as they were created at a time when no row index labels were available.